Rename pw.buttercup.mobile to com.buttercup (original)
