+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "pw.buttercup.mobile",
+    package = "com.buttercup",
 )
 
 android_resource(
     name = "res",
-    package = "pw.buttercup.mobile",
+    package = "com.buttercup",
     res = "src/main/res",
 )
